analyzers/volatility_analyzer/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.0
pytest==8.0.0
